Fixing problem with calibration filename
When calibration files are copied they retained their original
filename. However, the calibration system expeced a file named
'calibration' with the extension retained.

I've modified main.py to rename the file when copied so that the
name is as expected.

Testing:
Ran the software with an arbitrarily named calibration file and
verified that it was correctly renamed. Then verified that the
calibration process runs without error.