Started work on UI redesign.
Going for a simple main window which allows the user to select
their project directory using a file dialog and then run
the different tools as in the previous version.

I want this to default to whatever the last directory in use was
which will require some settings data to be stored in the data
directory.

I've started working on the UI layout but have yet to plumb
up the callbacks. I'm planning just to code these into the
custom widgets as it makes as much sense as any other option.