Mixed UI changeset
Added a lot of functionality to the new UI.

1. Small information frame added to display information about the
   current tool selection. This updates when the radio buttons are
   changed
2. Added graphical project selection.
3. Added project caching and global params functionality
4. Added new project creation

'New project' relies on a project file. At present the
project file doesn't store any information but just acts as
a flag to designate the parent directory as a DTrack directory.
There's no reason it can't store some info though. I have a plan
to implement project files as a rolling change which can be cemented
after testing.

There are some minor UI issues which can be fixed later.