Updated centroid computation to avoid failure if thresholding fails
Centroid now defaults to the centre of the bounding box if
thresholding of the beetle fails. I've also changed bbox selection
so that you select from the centre (which makes it easier to
ensure the beetle is at the centre).

This isn't a perfect solution and I want to try using an alternative
colour space or chromaticity coordinates to fix this. At least now
it won't crash...