Linked calibration mechanism into the new UI and updated chessboard selector
Calibration can now be called from the new UI, this is a breaking change
as it changes the calibration function call.

The chessboard has also been updated. The displayed pattern was wrong,
because I misremembered how OpenCV chessboard detection worked.
The full pattern is now displayed with the pattern detected by OpenCV
layered over the top.