Started building a new calibration script
Started a new calibration script based on randomly selecting
frames from the calibration video in which the chessboard
could be found; extrinsic calibration frames have to be specified
manually.

The script then builds a cache of the images in which the
checkerboard was found and the coordinates of the corners.

This could be extended later on to allow a user to specify their
own image files, or using the video pause/select method we
currently employ.