Addressing undocumented bug found by Sherry (caused by issue #17)
Calibration files would not be stored due to a missing 'calib_data'
directory. This is (in part) caused by the brittle directory handling
discussed in issue #17. Have added an existance check for the
directory and re-written all of the file-handling code to use
os.path.join instead of string concatenation.