Added (apparently) working scaling code
Have added code which estimates the scale parameter to transform
pixel distances in a calibrated image into millimetres.

I have only tested this on one calibration image so there are likely
problems and some brittle behaviour which we can iron out as we go.

This is still only a test script but this can now be worked into the
main software.