Added a basic calibration manager window with some backend features
At the moment the calibration manager will check whether a calibration
exists in the right place and inform the user (calibration found, not found,
or rudimentary corruption checking).

Checking for a calibration file necessitated adding more parameters
to the project file and strictly setting the calibration cache directory
as project_directory/calibration_cache. This is not planned to be
configurable.