Updated default handling in dtrack_params
An issue was found whereby a user had a param file but the elements
were null (which corresponds to None when loaded). This caused problems
with the project file defaults which assume that the param file
elements are strings.

Have updated the param file to either set default values in the
param file ("") or check an existing file for null values. This
means that if no project file exists, or null values are found
then they'll be replaced by the empty string which may cause problems
if the user tries to do anything without first choosing their project.