Overhauling calibration and calibration verification
Autocalibration included two perspective transformations. One
for extrinsic calibration, then a second which brought remapped
the image back into displayable space. This second transformation
introduced distortion in the frame which messed up distance measurements.
This has been corrected by removing the transform and computing a
(rough) translation transformation to move the centre of the chessboard
to the centre of the image. This has improved the calibration remarkably.

Calibration verification has been improved to include a virtual arena.
The user supplies four points which lie on the arena. These are
used to work out where the arena centre should be. A virtual
arena is then drawn on the frame which allows the user to verify
that the image has been properly calibrated (that the image is circular).

Tool frame has been updated with respect to functional changes.

The arena drawn in plots is now 50cm instead of 35 (though I suspect
this is wrong).

The readme has been updated to reflect the changes in the calibration
verification process.