Updated check calibration to display full transformed frame.
The old check calibration method assumed the chessboard should be
the centre of the frame which may not be the case. The new version
will check where the transformed image corners go, compute a bounding
box (the new frame) then apply a translation to the image so that
the bounding box starts at (0,0).

This makes for a nicer calibration verification tool as all of the
image pixels will always be visible but also means we may be
able to show the tracks overlayed on the calibration frame.