Created an options panel with options written back to the params/project file
Have created a new window with configuration options both for
DTrack2 and the current project.

Have updated default handling in the params file to make sure
all elements are of the correct type.

Have also modified the write-out process for the project and
params files in order to make them human readable.