Removed project options tab
The configuration tool had a tab for project options and dtrack2
options. For the moment it seems more sensible just to configure
these settings globally as required because each 'project' is
one beetle (i.e. settings likely to be set for a series of
projects).

I've retained the track point setting in the project file as
this still makes some sense, however it's coupled and identical to
the dtrack_params track point setting. I think there will still
be scope for having project settings down the line but for now
there is this code redundancy.