Added default video directory functionality to video selector
Linked up the default video directory setting with the video selector
tool so that video selection now opens in the default directory selected
by the user. Also checked that the default option "." works as expected
(opening in the same directory as all the tool files).

Also noticed and fixed a minor bug in the SelectorFrame class where
cancelling the file selection resulted in the Entry being cleared.
The code didn't check for user cancellation so this has been added
and tested.