Added an option for tracking interval
Added an option to track every nth frame as opposed to tracking
every individual frame in the source video.

The changes include changes to the configuration screen, params file,
autotracker, and track processing.

Critically, time tracking has now changed to use the millisecond
timestamps read from each frame by OpenCV. Timestamps are now stored
in a separate CSV.

Track processing now takes account of the fact that this file
may or may not exist and also may be incomplete for backwards
compatibility purposes.

Some code has also been added to allow the autotracker to carry
the initially specified bounding box into each subsequent track
which reduces the amount of interaction required. This is not
yet exposed to the user as there needs to be a way to undo the
current selection.