Adding feature #24
Have now added an option to store an ROI over multiple tracks so
that the user doesn't need to specify it every time. This can
also be re-specified if required.

README as has been updated to include a short description of the
option and some instructions are also included on the video frame
in the autotracker (if the option is enabled).