Removed autotracker track point information from project file
This was included as a 'just in case' thing but it's not used and
causes a problem whereby trying to update the software settings
will cause an error if there is no project selected.

The project file should now contain no software configuration
data (or at least nothing which is in use).