Fix new project cancellation bug
When selecting a new project, if the user clicks 'Cancel' the
filedialog would return an empty stringinstead of an empty tuple.
This bypassed the cancellation check and attempted to create a new
project directory.

This has now been fixed by adding a check for the empty string
or an empty tuple. Empty tuples are the standard check value
so this may be a problem with other file dialogs too!