Fix bbox state bug in autotracker
If user cancelled the first ROI selection this would cause the
software to crash because the value of first bbox wasn't correctly
checked. This has now been fixed. Behaviour is now that cancelling
selection returns you to the video and trying to track automatically
asks for ROI selection.

Similarly, if the user wanted to re-define the bbox but cancelled,
this would overwrite the old bbox without providing anything new.
Behaviour is now that cancelling restores the old bbox.