Adding axes flip options to track plotting.
The perspective on the tracks can be changed by the conversion from
OpenCV image coordinates to plot coordinates. The perspective can
also be changed by calibration. This commit adds two checkboxes
to the options screen which allows the user to mirror the x/y axes
as they see fit. The mirroring is applied as the data is plotted
and no intermediate form is stored.